requirements.txt
langchain
langchain-core
langchain-chroma
langchain-ollama

chromadb

unstructured
unstructured[pdf]

python-dotenv

streamlit

pillow
opencv-python
pytesseract

numpy
